+++ tools/perf/ucc_pt_coll_reduce.cc
@@ -87,11 +87,11 @@ void ucc_pt_coll_reduce::free_args(ucc_pt_test_args_t &test_args)
     }
 }
 
-float ucc_pt_coll_reduce::get_bw(float time_ms, int grsize,
+double ucc_pt_coll_reduce::get_bw(float time_ms, int grsize,
                                  ucc_pt_test_args_t test_args)
 {
     ucc_coll_args_t &args = test_args.coll_args;
-    float            S    = args.src.info.count *
+    double           S    = args.src.info.count *
                             ucc_dt_size(args.src.info.datatype);
 
     return S / time_ms / 1000.0;
